refiguring admin
